feat(typescript): Upgrad to typescript 2.0, extract execeptions to common to stop browser linking server modules, remove typings, remove
BREAKING CHANGE:
You typescript 2.0 lands with this release, with it comes a total refactor on how tsconfig is used. To see the recommended implementation, review the changes in https://github.com/ubiquits/quickstart/pull/22
