--- common.d.ts
@@ -1,1 +1,0 @@
-export * from './lib/common/index';
