+++ package.json
@@ -9,7 +9,6 @@
     "test": "u test -s",
     "coveralls": "u coveralls",
     "build": "u build",
-    "postinstall": "typings install",
     "postpublish": "u deploy docs"
   },
   "repository": {
@@ -53,10 +52,21 @@
     "timestamp": "0.0.1",
     "typeorm": "0.0.2-alpha.43",
     "vantage": "^1.7.0",
-    "zone.js": "^0.6.12"
-  },
-  "devDependencies": {
-    "@ubiquits/toolchain": "^0.1.42",
+    "zone.js": "^0.6.12",
+    "@types/chalk": "^0.4.28",
+    "@types/chance": "^0.7.28",
+    "@types/dotenv": "^2.0.17",
+    "@types/express": "^4.0.29",
+    "@types/express-serve-static-core": "^4.0.29",
+    "@types/hapi": "^13.0.28",
+    "@types/jasmine": "^2.2.29",
+    "@types/lodash": "0.0.27",
+    "@types/moment": "^2.11.28",
+    "@types/node": "^4.0.29",
+    "@types/proxyquire": "^1.3.26",
+    "@types/socket.io": "^1.4.26",
+    "@types/validator": "^4.5.26",
+    "@ubiquits/toolchain": "^0.2.0",
     "proxyquire": "^1.7.10"
   },
   "directories": {
