+++ src/common/exeptions/index.ts
@@ -1,5 +1,5 @@
 /**
- * @module server
+ * @module common
  */
 /** End Typedoc Module Declaration */
 export * from './exceptions';
