+++ src/common/index.ts
@@ -19,3 +19,4 @@ export * from './stores';
 export * from './validation';
 export * from './services';
 export * from './registry';
+export * from './exeptions';
