+++ src/common/services/logger.service.ts
@@ -4,7 +4,6 @@
 /** End Typedoc Module Declaration */
 import { Injectable } from '@angular/core';
 import { AbstractService } from './service';
-import { InternalServerErrorException } from '../../server/exeptions/exceptions';
 
 /**
  * Syslog Levels
