+++ src/common/stores/mock.store.ts
@@ -9,8 +9,6 @@ import { Injector } from '@angular/core';
 import { Chance } from 'chance';
 import * as _ from 'lodash';
 
-export type ChanceInstance = Chance.Chance;
-
 /**
  * Provides abstract class to build concrete mock stores which can create new mock instances of
  * models.
@@ -21,7 +19,7 @@ export abstract class MockStore<T extends AbstractModel> extends AbstractStore<T
    * Instance of chancejs
    * @see http://chancejs.com
    */
-  protected chanceInstance: ChanceInstance;
+  protected chanceInstance: Chance.Chance;
 
   protected modelCollection: Collection<T>;
 
@@ -45,7 +43,7 @@ export abstract class MockStore<T extends AbstractModel> extends AbstractStore<T
    * @param seed
    * @returns {ChanceInstance}
    */
-  protected chance(seed?: any): ChanceInstance {
+  protected chance(seed?: any): Chance.Chance {
     if (!this.chanceInstance || !!seed) {
       this.chanceInstance = new Chance(seed);
     }
