+++ src/server/controllers/abstract.controller.spec.ts
@@ -10,7 +10,7 @@ import { Request } from './request';
 import { Response } from './response';
 import { AbstractController } from './abstract.controller';
 import { Route } from './route.decorator';
-import { UnavailableForLegalReasonsException } from '../exeptions/exceptions';
+import { UnavailableForLegalReasonsException } from '../../common/exeptions/exceptions';
 
 @Injectable()
 class TestController extends AbstractController {
