+++ src/server/stores/db.store.ts
@@ -8,8 +8,8 @@ import { Database } from '../services/database.service';
 import { Logger } from '../../common/services/logger.service';
 import { AbstractStore, Query } from '../../common/stores/store';
 import { Collection } from '../../common/models/collection';
-import { NotFoundException } from '../exeptions/exceptions';
 import { Repository, Connection } from 'typeorm';
+import { NotFoundException } from '../../common/exeptions/exceptions';
 
 /**
  * Database store should be extended with a specific implementation for a model. Interacts with
