--- typings.json
@@ -1,21 +1,0 @@
-{
-  "globalDependencies": {
-    "chance": "registry:dt/chance#0.7.3+20160317120654",
-    "dotenv": "registry:dt/dotenv#2.0.0+20160327131627",
-    "express": "registry:dt/express#4.0.0+20160317120654",
-    "express-serve-static-core": "registry:dt/express-serve-static-core#0.0.0+20160701212727",
-    "hapi": "registry:dt/hapi#13.0.0+20160602125023",
-    "jasmine": "registry:dt/jasmine#2.2.0+20160621224255",
-    "moment-node": "registry:dt/moment-node#2.11.1+20160511043338",
-    "node": "registry:dt/node#6.0.0+20160621231320",
-    "serve-static": "registry:dt/serve-static#0.0.0+20160606155157",
-    "socket.io": "registry:dt/socket.io#1.4.4+20160317120654",
-    "validator": "registry:dt/validator#4.5.1+20160316155526"
-  },
-  "dependencies": {
-    "chalk": "registry:npm/chalk#1.0.0+20160211003958",
-    "lodash": "registry:npm/lodash#4.0.0+20160701205107",
-    "mime": "registry:npm/mime#1.3.0+20160423043021",
-    "proxyquire": "registry:npm/proxyquire#1.0.0+20160211003958"
-  }
-}
